added statistic_class.py and class Statistic with methods for statistics
