+++ infra/main.tf
@@ -51,7 +51,7 @@ module "lambda_dynamodb_stream" {
   dynamodb_table_name = module.dynamodb.dynamodb_table_name
   lambda_role         = module.iam.iam_role_arn
   file_hash           = var.file_hash
-  handler_name        = "app.dynamodb_stream.handler"
+  handler_name        = "handlers.mail_pet_stream.handler"
   function_suffix     = "dynamodb_stream"
   s3_bucket_artifacts = var.s3_bucket_artifacts
   env_variables = {
@@ -73,7 +73,7 @@ module "lambda_statistics_weekly" {
   dynamodb_table_name = module.dynamodb.dynamodb_table_name
   lambda_role         = module.iam.iam_role_arn
   file_hash           = var.file_hash
-  handler_name        = "app.statistics.handler"
+  handler_name        = "handlers.mail_statistics.handler"
   function_suffix     = "weekly_statistics"
   env_variables = {
     s3_bucket_name : module.photo-s3bucket.s3_bucket_name,
@@ -89,7 +89,7 @@ module "lambda_statistics_overall" {
   dynamodb_table_name = module.dynamodb.dynamodb_table_name
   lambda_role         = module.iam.iam_role_arn
   file_hash           = var.file_hash
-  handler_name        = "app.statistics.handler"
+  handler_name        = "handlers.mail_statistics.handler"
   function_suffix     = "overall_statistics"
   env_variables = {
     s3_bucket_name : module.photo-s3bucket.s3_bucket_name,
